# claims-validation-api/generated/src/main/java/uk/gov/justice/laa/dstew/payments/claims/api/ValidationApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.18.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package uk.gov.justice.laa.dstew.payments.claims.api;

import uk.gov.justice.laa.dstew.payments.claims.model.ClaimValidationRequest;
import uk.gov.justice.laa.dstew.payments.claims.model.ValidationResult;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-03-13T21:51:30.435742Z[Europe/London]", comments = "Generator version: 7.18.0")
@Validated
@Tag(name = "validation", description = "the validation API")
public interface ValidationApi {

    String PATH_VALIDATE_CLAIM = "/v1/validation/claim";
    /**
     * POST /v1/validation/claim : Validate a claim
     *
     * @param claimValidationRequest Claim validation request (required)
     * @return Validation completed (status code 200)
     *         or Bad request (status code 400)
     *         or Unauthorized (status code 401)
     *         or Forbidden (status code 403)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "validateClaim",
        summary = "Validate a claim",
        tags = { "validation" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Validation completed", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ValidationResult.class))
            }),
            @ApiResponse(responseCode = "400", description = "Bad request"),
            @ApiResponse(responseCode = "401", description = "Unauthorized"),
            @ApiResponse(responseCode = "403", description = "Forbidden"),
            @ApiResponse(responseCode = "500", description = "Internal server error")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = ValidationApi.PATH_VALIDATE_CLAIM,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    ResponseEntity<ValidationResult> validateClaim(
        @Parameter(name = "ClaimValidationRequest", description = "Claim validation request", required = true) @Valid @RequestBody ClaimValidationRequest claimValidationRequest
    );

}